+++ .testassets/basictest.py
@@ -5,39 +5,44 @@ from duality import Record
 
 t = Record()
 
-class Queue:
+class App:
 
+    @t.entry(option_name='init', autoinit=True)
     def __init__(self, dictionary = {}):
         self.dictionary = dictionary
         return
 
     @t.entry(option_name='zbrajanje')
     def zbrajanje(self):
-        x = int(input('Unesi prvi broj: '))
-        y = int(input('Unesi drugi broj: '))
+        x = t.define(input_val = 'x', dtype = 'int')
+        y = t.define(input_val = 'y', dtype = 'int')
         rez = print('Zbroj je: ', x + y)
         return rez
 
     @t.entry(option_name='oduzimanje')
     def oduzimanje(self):
-        x = int(input('Unesi prvi broj: '))
-        y = int(input('Unesi drugi broj: '))
+        x = t.define(input_val = 'x', dtype = 'int')
+        y = t.define(input_val = 'y', dtype = 'int')
         rez = print('Razlika je: ', x - y)
         return rez
 
     @t.entry(option_name='množenje')
     def množenje(self):
-        x = int(input('Unesi prvi broj: '))
-        y = int(input('Unesi drugi broj: '))
+        x = t.define(input_val = 'x', dtype = 'int')
+        y = t.define(input_val = 'y', dtype = 'int')
         rez = print('Umnožak je: ', x * y)
         return rez
 
     @t.entry(option_name='dijeljenje')
     def dijeljenje(self):
-        x = int(input('Unesi prvi broj: '))
-        y = int(input('Unesi drugi broj: '))
+        x = t.define(input_val = 'x', dtype = 'int')
+        y = t.define(input_val = 'y', dtype = 'int')
         rez = print('Produkt je: ', x // y)
         return rez
 
+    def initconfig(self):
+        t.config(type = 'dynamic', queue=True)
 
-t.config(queue=True)
+a = App()
+
+a.initconfig()