+++ duality/plugins/__init__.py
@@ -24,6 +24,7 @@ from vandal.plugins.types import (
     NumberArrayAlike,
     AnyArrayAlike,
     AnyVectorAlike,
+    AnyType,
 )
 
 # all relevant contents.
@@ -48,6 +49,7 @@ __all__ = [
     'NumberArrayAlike',
     'AnyArrayAlike',
     'AnyVectorAlike',
+    'AnyType',
 ]
 
 # all available types.
@@ -72,4 +74,5 @@ __types__ = [
     NumberArrayAlike,
     AnyArrayAlike,
     AnyVectorAlike,
+    AnyType,
 ]